Added ImageLoader to cache and download image
